Use one story file for all working group page tabs
